Move channel withdraw verification to its corresponding module.
